# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Api",
        "description": "Lumen Swagger Api",
        "license": {
            "name": "my api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Api Documentations my api"
        }
    ],
    "paths": {
...
        "/place": {
            "get": {
                "tags": [
                    "Place"
                ],
                "summary": "List place",
                "description": "Returns data of places",
                "operationId": "List",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Place"
                ],
                "summary": "Store place",
                "operationId": "Store",
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de el lugar",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripcion del lugar",
                                        "type": "string"
                                    },
                                    "x": {
                                        "description": "Longitud de posición",
                                        "type": "number"
                                    },
                                    "y": {
                                        "description": "Latitud de posición",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
        "/place/{id}": {
            "get": {
                "tags": [
                    "Place"
                ],
                "summary": "Show place",
                "description": "Returns data of place",
                "operationId": "Show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "place id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Place"
                ],
                "summary": "Update place",
                "operationId": "Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "place id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de el lugar",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripcion del lugar",
                                        "type": "string"
                                    },
                                    "x": {
                                        "description": "Longitud de posición",
                                        "type": "number"
                                    },
                                    "y": {
                                        "description": "Latitud de posición",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Place"
                ],
                "summary": "Delete existing place",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteplace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "place id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "authorization": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}